chore(builder): using unified version of webpack-chain
