+++ .changeset/tasty-news-sparkle.md
@@ -1,0 +1,13 @@
+---
+'@modern-js/builder-webpack-provider': patch
+'@modern-js/builder-shared': patch
+'@modern-js/plugin-data-loader': patch
+'@modern-js/plugin-garfish': patch
+'@modern-js/plugin-storybook': patch
+'@modern-js/plugin-bff': patch
+'@modern-js/utils': patch
+---
+
+chore(builder): using unified version of webpack-chain
+
+chore(builder): 使用统一的 webpack-chain 版本
