--- packages/builder/builder-shared/compiled/webpack-5-chain/package.json
@@ -1,1 +1,0 @@
-{"name":"webpack-5-chain","author":"Eli Perelman <eli@eliperelman.com>","version":"8.0.1","license":"MPL-2.0","typings":"types/index.d.ts"}
