+++ packages/builder/builder-shared/package.json
@@ -50,10 +50,6 @@
       "default": "./compiled/webpack-bundle-analyzer/index.js"
     },
     "./babel-loader": "./compiled/babel-loader/index.js",
-    "./webpack-5-chain": {
-      "types": "./compiled/webpack-5-chain/types/index.d.ts",
-      "default": "./compiled/webpack-5-chain/index.js"
-    },
     "./pug": {
       "types": "./compiled/pug/index.d.ts",
       "default": "./compiled/pug/index.js"
@@ -89,9 +85,6 @@
       ],
       "webpack-bundle-analyzer": [
         "./compiled/webpack-bundle-analyzer/index.d.ts"
-      ],
-      "webpack-5-chain": [
-        "./compiled/webpack-5-chain/types/index.d.ts"
       ],
       "test-stub": [
         "./dist/test-stub/index.d.ts",
@@ -158,7 +151,6 @@
       "./open": "./compiled/open/index.js",
       "./babel-loader": "./compiled/babel-loader/index.js",
       "./webpack-bundle-analyzer": "./compiled/webpack-bundle-analyzer/index.js",
-      "./webpack-5-chain": "./compiled/webpack-5-chain/index.js",
       "./test-stub": "./dist/test-stub/index.js",
       "./zod": "./dist/exports/zod.js",
       "./fork-ts-checker-webpack-plugin": "./dist/exports/fork-ts-checker-webpack-plugin.js",
