+++ packages/cli/plugin-bff/package.json
@@ -85,8 +85,7 @@
     "memfs": "^3.3.0",
     "ts-jest": "^29.1.0",
     "typescript": "^5",
-    "webpack": "^5.88.1",
-    "webpack-chain": "^6.5.1"
+    "webpack": "^5.88.1"
   },
   "sideEffects": false,
   "publishConfig": {
