+++ packages/cli/plugin-data-loader/package.json
@@ -84,7 +84,6 @@
     "ts-jest": "^29.1.0",
     "typescript": "^5",
     "webpack": "^5.88.1",
-    "webpack-chain": "^6.5.1",
     "react": "^18",
     "react-dom": "^18"
   },
