+++ packages/cli/plugin-storybook/package.json
@@ -86,8 +86,7 @@
     "react-dom": "^17.0.2",
     "require-from-string": "^2.0.2",
     "typescript": "^5",
-    "webpack": "^5.88.1",
-    "webpack-chain": "^6.5.1"
+    "webpack": "^5.88.1"
   },
   "peerDependencies": {
     "@modern-js/runtime": "workspace:^2.31.2",
