+++ packages/runtime/plugin-garfish/package.json
@@ -103,8 +103,7 @@
     "react": "^18",
     "react-dom": "^18",
     "react-router-dom": "^6.8.1",
-    "typescript": "^5",
-    "webpack-chain": "^6.5.1"
+    "typescript": "^5"
   },
   "sideEffects": false,
   "publishConfig": {
