+++ packages/toolkit/utils/compiled/webpack-chain/package.json
@@ -1,1 +1,1 @@
-{"name":"webpack-chain","author":"Eli Perelman <eli@eliperelman.com>","version":"6.5.1","license":"MPL-2.0","typings":"types/index.d.ts"}
+{"name":"webpack-chain","author":"Eli Perelman <eli@eliperelman.com>","version":"8.0.1","license":"MPL-2.0","typings":"types/index.d.ts"}
