+++ scripts/prebundle/package.json
@@ -134,9 +134,8 @@
     "url-join": "4.0.1",
     "url-loader": "4.1.1",
     "v8-compile-cache": "2.3.0",
-    "webpack-5-chain": "8.0.1",
+    "webpack-chain": "npm:webpack-5-chain@8.0.1",
     "webpack-bundle-analyzer": "4.5.0",
-    "webpack-chain": "6.5.1",
     "webpack-dev-middleware": "6.0.1",
     "webpack-manifest-plugin": "5.0.0",
     "webpack-merge": "5.8.0",
